+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Allow `unpack` of `Tuples` by implementing `IteratorAggregate`
 
 ## 3.4.2 - 2018-06-28
 - Fix `static creators` which are `deprecated` in some situation to allow `override` them